[refactor] lobby page 컴포넌트 분리
